+++ src/Application/Playback/PlayApplication.php
@@ -95,7 +95,7 @@ final class PlayApplication implements ApplicationInterface
             $action = 'stream';
         }
         // allow disabling stat recording from the play url
-        if ($cache === '1' || !in_array($type, array('song', 'video'))) {
+        if ($cache === '1' || !in_array($type, array('song', 'video', 'podcast_episode'))) {
             debug_event('play/index', 'record_stats disabled: cache {' . $type . "}", 5);
             $record_stats = false;
         }