+++ src/Module/Api/Json_Data.php
@@ -354,6 +354,7 @@ class Json_Data
                 "averagerating" => ($rating->get_average_rating() ?: null),
                 "mbid" => $artist->mbid,
                 "summary" => $artist->summary,
+                "time" => $artist->time,
                 "yearformed" => $artist->yearformed,
                 "placeformed" => $artist->placeformed
             ));
@@ -430,6 +431,7 @@ class Json_Data
                 $disk = (count($album->album_suite) <= 1) ? $album->disk : count($album->album_suite);
             }
 
+            $theArray['time']          = (int) $album->total_duration;
             $theArray['year']          = (int) $album->year;
             $theArray['tracks']        = $songs;
             $theArray['disk']          = (int) $disk;
@@ -577,6 +579,46 @@ class Json_Data
 
         return json_encode(array("share" => $allShares), JSON_PRETTY_PRINT);
     } // shares
+
+    /**
+     * bookmarks
+     *
+     * This returns bookmarks to the user, in a pretty json document with the information
+     *
+     * @param array $bookmarks (description here...)
+     * @return string return JSON
+     */
+    public static function bookmarks($bookmarks)
+    {
+        if ((count($bookmarks) > self::$limit || self::$offset > 0) && self::$limit) {
+            $bookmarks = array_splice($bookmarks, self::$offset, self::$limit);
+        }
+
+        $allBookmarks = [];
+        foreach ($bookmarks as $bookmark_id) {
+            $bookmark = new Bookmark($bookmark_id);
+            $bookmark->format();
+            $bookmark_user          = $bookmark->f_user;
+            $bookmark_object_type   = $bookmark->object_type;
+            $bookmark_object_id     = $bookmark->object_id;
+            $bookmark_position      = $bookmark->position;
+            $bookmark_comment       = $bookmark->comment;
+            $bookmark_creation_date = $bookmark->creation_date;
+            $bookmark_update_date   = $bookmark->update_date;
+            // Build this element
+            array_push($allBookmarks, [
+                "id" => (string) $bookmark_id,
+                "owner" => $bookmark_user,
+                "object_type" => $bookmark_object_type,
+                "object_id" => $bookmark_object_id,
+                "position" => $bookmark_position,
+                "client" => $bookmark_comment,
+                "creation_date" => $bookmark_creation_date,
+                "update_date" => $bookmark_update_date]);
+        } // end foreach
+
+        return json_encode(array("bookmark" => $allBookmarks), JSON_PRETTY_PRINT);
+    } // bookmarks
 
     /**
      * catalogs