+++ src/Module/Api/Method/ShareDeleteMethod.php
@@ -42,7 +42,6 @@ final class ShareDeleteMethod
 
     /**
      * share_delete
-     *
      * MINIMUM_API_VERSION=420000
      *
      * Delete an existing share.