+++ src/Module/Api/Subsonic_Api.php
@@ -2345,9 +2345,10 @@ class Subsonic_Api
      */
     public static function getbookmarks($input)
     {
+        $user_id   = User::get_from_username($input['u'])->id;
         $response  = Subsonic_Xml_Data::createSuccessResponse('getbookmarks');
-        $bookmarks = Bookmark::get_bookmarks();
-        Subsonic_Xml_Data::addBookmarks($response, $bookmarks);
+        $bookmarks = Bookmark::get_bookmarks($user_id);
+        Subsonic_XML_Data::addBookmarks($response, $bookmarks);
         self::apiOutput($input, $response);
     }